refactor game logic into its own module
