+++ Cargo.toml
@@ -10,3 +10,13 @@ edition = "2018"
 strum = "0.22"
 strum_macros = "0.22"
 rand = "0.8"
+
+actix = "0.13"
+actix-files = "0.6"
+actix-web = "4"
+actix-web-actors = "4.1"
+
+env_logger = "0.9"
+log = "0.4"
+serde = "1"
+serde_json = "1"
