+++ src/logic/mod.rs
@@ -1,0 +1,4 @@
+pub mod game;
+mod card;
+
+pub use game::Game;
